+++ CHANGELOG.md
@@ -1,3 +1,61 @@
+## [1.4.1-beta.0](https://github.com/alexneo2003/playwright-azure-reporter/compare/v1.4.0-beta.0...v1.4.1-beta.0) (2022-12-30)
+
+
+### Bug Fixes
+
+* fixing some throwe error ([733e759](https://github.com/alexneo2003/playwright-azure-reporter/commit/733e759f82bf7dc17176af7f3ef04ce43593ceeb))
+
+
+
+# [1.4.0-beta.0](https://github.com/alexneo2003/playwright-azure-reporter/compare/v1.3.1-beta.0...v1.4.0-beta.0) (2022-12-29)
+
+
+### Features
+
+* publishTestResultsMode config to set mode when publish test results ([4e25019](https://github.com/alexneo2003/playwright-azure-reporter/commit/4e2501926beb4680c4ef770273847fae39c57b71))
+
+
+
+## [1.3.1-beta.0](https://github.com/alexneo2003/playwright-azure-reporter/compare/v1.3.0-beta.0...v1.3.1-beta.0) (2022-12-22)
+
+
+
+# [1.3.0-beta.0](https://github.com/alexneo2003/playwright-azure-reporter/compare/v1.2.0-beta.2...v1.3.0-beta.0) (2022-10-31)
+
+
+### Features
+
+* define multiple test cases for a single test ([a89847a](https://github.com/alexneo2003/playwright-azure-reporter/commit/a89847acd8a4115cfb54329b27c7941ac403a792))
+
+
+
+# [1.2.0-beta.2](https://github.com/alexneo2003/playwright-azure-reporter/compare/v1.2.0-beta.1...v1.2.0-beta.2) (2022-10-14)
+
+
+### Bug Fixes
+
+* configurationIds implementing ([9c9ad44](https://github.com/alexneo2003/playwright-azure-reporter/commit/9c9ad44b4a798e336b534f46a804e64d9c8cc31a))
+
+
+
+# [1.2.0-beta.1](https://github.com/alexneo2003/playwright-azure-reporter/compare/v1.2.0-beta.0...v1.2.0-beta.1) (2022-10-13)
+
+
+### Bug Fixes
+
+* optional configurationIds ([ffd01c3](https://github.com/alexneo2003/playwright-azure-reporter/commit/ffd01c34ea5a47c0075364ec286f46c233baa6f8))
+
+
+
+# [1.2.0-beta.0](https://github.com/alexneo2003/playwright-azure-reporter/compare/v1.1.1...v1.2.0-beta.0) (2022-09-30)
+
+
+### Features
+
+* Refactoring ([8f2545d](https://github.com/alexneo2003/playwright-azure-reporter/commit/8f2545d79cf9e3015498fcc01823d9ecbb47de8a))
+
+
+
 ## [1.1.1](https://github.com/alexneo2003/playwright-azure-reporter/compare/v1.1.0...v1.1.1) (2022-07-26)